crazy experimental/debugging code.
creates a modified event queue for buffer_id() once again
calculates md5s based on what should be in the file and what is actually there, and what we see immediately thereafter.
makes a queue for patches if a modified notification is pending.
once we see the notification, we empty the queue (in the next tick)
we also empty the queue if we get anothe rpatch on the line